+++ packages/sdk/src/subscriber.ts
@@ -17,9 +17,10 @@ export default class ConnectionSubscriber extends ConnectionBase {
    */
   // biome-ignore lint/suspicious/noConfusingVoidType: stream が <MediaStream | void> なのでどうしようもない
   async connect(): Promise<MediaStream | void> {
-    if (this.options.multistream) {
-      await Promise.race([
-        this.multiStream().finally(() => {
+    // options.multistream が明示的に false を指定した時だけレガシーストリームにする
+    if (this.options.multistream === false) {
+      const stream = await Promise.race([
+        this.legacyStream().finally(() => {
           this.clearConnectionTimeout()
           this.clearMonitorSignalingWebSocketEvent()
         }),
@@ -28,10 +29,10 @@ export default class ConnectionSubscriber extends ConnectionBase {
       ])
       this.monitorWebSocketEvent()
       this.monitorPeerConnectionState()
-      return
+      return stream
     }
-    const stream = await Promise.race([
-      this.singleStream().finally(() => {
+    await Promise.race([
+      this.multiStream().finally(() => {
         this.clearConnectionTimeout()
         this.clearMonitorSignalingWebSocketEvent()
       }),
@@ -40,13 +41,12 @@ export default class ConnectionSubscriber extends ConnectionBase {
     ])
     this.monitorWebSocketEvent()
     this.monitorPeerConnectionState()
-    return stream
   }
 
   /**
-   * シングルストリームで Sora へ接続するメソッド
+   * レガシーストリームで Sora へ接続するメソッド
    */
-  private async singleStream(): Promise<MediaStream> {
+  private async legacyStream(): Promise<MediaStream> {
     await this.disconnect()
     this.setupE2EE()
     const ws = await this.getSignalingWebSocket(this.signalingUrlCandidates)